add middleware

+++ routes/web.php
@@ -10,19 +10,19 @@ Route::get('/', [JobController::class, 'homepage'])->name('homepage');
 
 
 // Routes for User model
-Route::get('/register', [UserController::class, 'registrationForm'])->name('userRegistrationPage');
-Route::post('/register', [UserController::class, 'registerUser'])->name('registerUser');
+Route::get('/register', [UserController::class, 'registrationForm'])->name('userRegistrationPage')->middleware('guest');
+Route::post('/register', [UserController::class, 'registerUser'])->name('registerUser')->middleware('guest');
 
-Route::get("/login", [UserController::class, 'loginForm'])->name('loginFormPage');
-Route::post("/login", [UserController::class, 'loginUser'])->name('loginUser');
-Route::get("/logout", [UserController::class, 'logout'])->name('logout');
+Route::get("/login", [UserController::class, 'loginForm'])->name('loginFormPage')->middleware('guest');
+Route::post("/login", [UserController::class, 'loginUser'])->name('loginUser')->middleware('guest');
+Route::get("/logout", [UserController::class, 'logout'])->name('logout')->middleware('guest');
 
 Route::get('/user/{id}', [UserController::class, 'userDetails'])->name('userDetails');
 
 
 // Routes for Job model
-Route::get('/post-job-ad', [JobController::class, 'createJobForm'])->name('jobFormPage');
-Route::post('/post-job-ad', [JobController::class, 'createJobAd'])->name('makeJob');
+Route::get('/post-job-ad', [JobController::class, 'createJobForm'])->name('jobFormPage')->middleware('auth');
+Route::post('/post-job-ad', [JobController::class, 'createJobAd'])->name('makeJob')->middleware('auth');
 
-Route::get("/post-job-ad/{id}", [JobController::class, 'jobDetailsPage'])->name('jobDetails');
-Route::delete("/post-job-ad/{id}", [JobController::class, 'deleteJob'])->name('deleteJob');
+Route::get("/post-job-ad/{id}", [JobController::class, 'jobDetailsPage'])->name('jobDetails')->middleware('auth');
+Route::delete("/post-job-ad/{id}", [JobController::class, 'deleteJob'])->name('deleteJob')->middleware('auth');
